+++ qiskit/ignis/measurement/discriminator/discriminators.py
@@ -12,7 +12,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 from abc import ABC, abstractmethod
-from numpy import ndarray
+import re
 from typing import Union, List
 from sklearn.preprocessing import StandardScaler
 
@@ -29,7 +29,7 @@ class BaseDiscriminationFitter(ABC):
     """
 
     def __init__(self, cal_results: Union[Result, List[Result]],
-                 qubit_mask: List[int], expected_states: List[str],
+                 qubit_mask: List[int], expected_states: List[str] = None,
                  standardize: bool = False,
                  schedules: Union[List[str], List[Schedule]] = None):
         """
@@ -50,13 +50,18 @@ class BaseDiscriminationFitter(ABC):
                 instead of the schedules. If schedules is None, then all the
                 schedules in cal_results are used.
         """
+        # Regex pattern used to identify calibration schedules
+        self._cal_pattern = r'cal_\d+$'
 
         # Use all results in cal_results if schedules is None
         if schedules is None:
-            schedules = self._get_schedules(cal_results)
+            schedules = self._get_schedules(cal_results, 0)
 
         # Dict of form schedule name: expected state
         self._expected_state = {}
+        if expected_states is None:
+            expected_states = self._get_expected_states(schedules)
+
         self._add_expected_states(expected_states, schedules)
 
         # Used to rescale the xdata qubit by qubit.
@@ -75,8 +80,8 @@ class BaseDiscriminationFitter(ABC):
             else:
                 self._backend_result_list.append(cal_results)
 
-        self._xdata = self.get_xdata(self._backend_result_list, schedules)
-        self._ydata = self.get_ydata(self._backend_result_list, schedules)
+        self._xdata = self.get_xdata(self._backend_result_list, 0, schedules)
+        self._ydata = self.get_ydata(self._backend_result_list, 0, schedules)
 
     def _add_ydata(self, schedule: Union[Schedule, str]):
         """
@@ -104,13 +109,13 @@ class BaseDiscriminationFitter(ABC):
             schedules (Union[List[Schedule], List[str]]):
         """
         if schedules is None:
-            schedules = self._get_schedules(result)
+            schedules = self._get_schedules(result, 0)
 
         self._backend_result_list.append(result)
         self._add_expected_states(expected_states, schedules)
         self._schedules.extend(schedules)
-        self._xdata = self.get_xdata(self._backend_result_list, schedules)
-        self._ydata = self.get_ydata(self._backend_result_list, schedules)
+        self._xdata = self.get_xdata(self._backend_result_list, 0, schedules)
+        self._ydata = self.get_ydata(self._backend_result_list, 0, schedules)
 
         if refit:
             self.fit()
@@ -138,26 +143,105 @@ class BaseDiscriminationFitter(ABC):
             expected_state = expected_states[idx]
             self._expected_state[name] = expected_state
 
-    @staticmethod
-    def _get_schedules(results: Union[Result, List[Result]]) -> List[str]:
+    def is_calibration(self, result_name: str) -> bool:
+        """
+        Identify if a name corresponds to a calibration name identified by
+        the regex pattern self._cal_pattern.
+
+        Args:
+            result_name (str): name of the result to be tested.
+
+        Returns (bool):
+            True if the name of the result indicates that it is a calibration
+            result.
+        """
+        return re.match(self._cal_pattern, result_name) is not None
+
+    def _get_schedules(self, results: Union[Result, List[Result]],
+                       schedule_type_to_get: int) -> List[str]:
         """
         Extracts the names of all Schedules in a Result or a list of Result.
+        Only the results with a name that matches self._cal_pattern are
+        returned.
 
         Args:
             results (Union[Result, List[Result]]): the results for which to
                 extract the names,
+            schedule_type_to_get (int): defines which schedule type to include
+                in the returned schedules.
+                type == 0: only calibration schedules.
+                type == 1: only non-calibration schedules.
+                type == 2: all schedules in the results.
         Returns (List[str]):
             The name of the schedules in results.
         """
-        schedules = []
         if isinstance(results, Result):
-            for res in results.results:
-                schedules.append(res.header.name)
+            results_list = [results]
         else:
-            for result in results:
-                schedules.extend([_.header.name for _ in result.results])
+            results_list = results
+
+        schedules = []
+        for res in results_list:
+            for result in res.results:
+                self._append_to_schedules(result.header.name,
+                                          schedule_type_to_get, schedules)
 
         return schedules
+
+    def _append_to_schedules(self, name: str, schedule_type: int,
+                             schedules: list):
+        """
+        Helper function to append schedule names.
+
+        Args:
+            name (str): name of the schedule that may be appended to schedules.
+            schedule_type (int): defines which schedule type to include
+                in the returned schedules.
+                type == 0: only calibration schedules.
+                type == 1: only non-calibration schedules.
+                type == 2: all schedules in the results.
+            schedules (list): a list of schedule names.
+        """
+        if schedule_type == 0:
+            if self.is_calibration(name):
+                schedules.append(name)
+        elif schedule_type == 1:
+            if not self.is_calibration(name):
+                schedules.append(name)
+        elif schedule_type == 2:
+            schedules.append(name)
+        else:
+            raise QiskitError('schedule_type must be either\n'
+                              '0: get only calibration schedules\n'
+                              '1: get only non-calibration schedules\n'
+                              '2: get all schedules in results.')
+
+    @staticmethod
+    def _get_expected_states(schedules: Union[List[Schedule], List[str]]) \
+            -> List[str]:
+        """
+        Get the names of the expected_states based on the schedule names by
+        replacing the substring 'cal_' in the name with ''. E.g. the name
+        'cal_01' becomes '01'.
+
+        Args:
+            schedules (Union[List[Schedule], List[str]]): a list of schedules
+            or a list of schedule names. These schedules are used to identify
+            the names of the expected states.
+
+        Returns (List[str]):
+            expected_states extracted from the schedules.
+        """
+        expected_states = []
+        for schedule in schedules:
+            if isinstance(schedules, Schedule):
+                name = schedule.name
+            else:
+                name = schedule
+
+            expected_states.append(name.replace('cal_', ''))
+
+        return expected_states
 
     @property
     def expected_states(self):
@@ -199,6 +283,7 @@ class BaseDiscriminationFitter(ABC):
 
     @abstractmethod
     def get_xdata(self, results: Union[Result, List[Result]],
+                  schedule_type_to_get: int,
                   schedules: Union[List[str], List[Schedule]] = None) \
             -> List[List[float]]:
         """
@@ -209,6 +294,11 @@ class BaseDiscriminationFitter(ABC):
                 used to retrieve the level 1 data. If result is a list of
                 Result then the first Result to return the data of schedule in
                 schedules is used.
+            schedule_type_to_get (int): use to specify if we should return data
+                corresponding to:
+                0: calibration data only
+                1: non-calibration data
+                2: both calibration and non-calibration data
             schedules (Union[List[str], List[Schedule]]): Either the names of
                 the schedules or the schedules themselves.
 
@@ -219,12 +309,18 @@ class BaseDiscriminationFitter(ABC):
 
     @abstractmethod
     def get_ydata(self, results: Union[Result, List[Result]],
+                  schedule_type_to_get: int,
                   schedules: Union[List[str], List[Schedule]] = None) \
             -> List[str]:
         """
         Args:
             results (Union[Result, List[Result]]): results for which to
                 retrieve the y data (i.e. expected states).
+            schedule_type_to_get (int): use to specify if we should return data
+                corresponding to:
+                0: calibration data only
+                1: non-calibration data
+                2: both calibration and non-calibration data
             schedules (Union[List[str], List[Schedule]]): the schedules for
                 which to get the y data.