+++ CHANGELOG.md
@@ -19,8 +19,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Pulse calibrations for single qubits (\#292)
-- Pulse Discriminator (\#238, \#278)
+- Accreditation (\#252)
+- Pulse calibrations for single qubits (\#292, \#302, \#303, \#304)
+- Pulse Discriminator (\#238, \#278, \#297)
 
 ### Deprecated